requirements.txt
python-telegram-bot[job-queue]>=20.0
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.28.0
python-dotenv>=1.0.0
